group switched to secp256k1
