+++ src/main/scala/sigmastate/interpreter/Interpreter.scala
@@ -7,6 +7,7 @@ import org.bitbucket.inkytonik.kiama.rewriting.Rewriter.{everywherebu, rule, str
 import org.bitbucket.inkytonik.kiama.rewriting.Strategy
 import org.bouncycastle.math.ec.ECPoint
 import org.bouncycastle.math.ec.custom.djb.Curve25519Point
+import org.bouncycastle.math.ec.custom.sec.SecP256K1Point
 import sigmastate.basics.DLogProtocol.{DLogInteractiveProver, FirstDLogProverMessage}
 import scorex.crypto.hash.Blake2b256
 import scorex.util.ScorexLogging
@@ -15,7 +16,7 @@ import sigmastate.Values._
 import sigmastate.eval.IRContext
 import sigmastate.interpreter.Interpreter.ScriptEnv
 import sigmastate.lang.Terms.ValueOps
-import sigmastate.basics.{BcDlogGroup, Curve25519, DiffieHellmanTupleInteractiveProver, FirstDiffieHellmanTupleProverMessage}
+import sigmastate.basics._
 import sigmastate.interpreter.Interpreter.VerificationResult
 import sigmastate.lang.exceptions.InterpreterException
 import sigmastate.serialization.ValueSerializer
@@ -24,48 +25,6 @@ import sigmastate.{SType, _}
 
 import scala.util.Try
 
-
-object CryptoConstants {
-  type EcPointType = Curve25519Point
-
-  val dlogGroup: BcDlogGroup[EcPointType] = Curve25519
-  lazy val secureRandom = dlogGroup.secureRandom
-
-  def secureRandomBytes(howMany: Int) = {
-    val bytes = new Array[Byte](howMany)
-    secureRandom.nextBytes(bytes)
-    bytes
-  }
-
-  /** Size of the binary representation of any group element (2 ^ groupSizeBits == <number of elements in a group>) */
-  val groupSizeBits: Int = 256
-
-  /** Number of bytes to represent any group element as byte array */
-  val groupSize: Int = 256 / 8 //32 bytes
-
-  /** Group order, i.e. number of elements in the group */
-  val groupOrder: BigInteger = dlogGroup.order
-
-  //size of challenge in Sigma protocols, in bits
-  //if this anything but 192, threshold won't work, because we have polynomials over GF(2^192) and no others
-  //so DO NOT change the value without implementing polynomials over GF(2^soundnessBits) first
-  //and changing code that calls on GF2_192 and GF2_192_Poly classes!!!
-  implicit val soundnessBits: Int = 192.ensuring(_ < groupSizeBits, "2^t < q condition is broken!")
-}
-
-object CryptoFunctions {
-  lazy val soundnessBytes: Int = CryptoConstants.soundnessBits / 8
-
-  def hashFn(input: Array[Byte]): Array[Byte] = {
-    Blake2b256.hash(input).take(soundnessBytes)
-  }
-
-  def showECPoint(p: ECPoint) = {
-    val rawX = p.getRawXCoord.toString.substring(0, 6)
-    val rawY = p.getRawYCoord.toString.substring(0, 6)
-    s"ECPoint($rawX,$rawY,...)"
-  }
-}
 
 trait Interpreter extends ScorexLogging {
 
@@ -230,6 +189,7 @@ trait Interpreter extends ScorexLogging {
              message: Array[Byte]): Try[VerificationResult] = {
     verify(Interpreter.emptyEnv, exp, context, SigSerializer.toBytes(proof), message)
   }
+
 }
 
 object Interpreter {
@@ -241,4 +201,5 @@ object Interpreter {
   val ScriptNameProp = "ScriptName"
 
   def error(msg: String) = throw new InterpreterException(msg)
+
 }
